circular database solved v7.0
