+++ project/auth.py
@@ -1,0 +1,66 @@
+from flask import Blueprint, render_template, redirect, url_for, request, flash
+from werkzeug.security import generate_password_hash, check_password_hash
+from flask_login import login_user, logout_user, login_required
+from .__init__ import db
+from .models import User
+
+auth = Blueprint('auth', __name__)
+
+
+@auth.route('/login', methods=['GET'])
+def login():
+    return render_template('login.html')
+
+
+@auth.route('/signup')
+def signup():
+    return render_template('signup.html')
+
+
+@auth.route('/signup', methods=['POST'])
+def signup_post():
+    email = request.form.get('email')
+    username = request.form.get('username')
+    password = request.form.get('password')
+
+    user = User.query.filter_by(email=email).first()
+
+    if user:
+        flash('Email address already exists!')
+        return redirect(url_for('auth.login'))
+
+    new_user = User(email=email, username=username, password=generate_password_hash(password, method='sha256'))
+    try:
+        db.session.add(new_user)
+        db.session.commit()
+    except:
+        return 'Failed to add the { new_user } on database'
+    return redirect(url_for('auth.login'))
+
+
+@auth.route('/login', methods=['POST'])
+def login_post():
+    email = request.form.get('email')
+    password = request.form.get('password')
+    remember = True if request.form.get('remember') else False
+
+    user = User.query.filter_by(email=email).first()
+
+    if not user or not check_password_hash(user.password, password):
+        flash('Please check your login details and try again.')
+        return redirect(url_for('auth.login'))
+
+    login_user(user, remember=remember)
+    return redirect(url_for('main.profile'))
+
+
+@auth.route('/logout')
+@login_required
+def logout():
+    logout_user()
+    return redirect(url_for('main.index'))
+
+
+@auth.route('/contact')
+def contact():
+    return render_template('contact.html')
