+++ project/templates/base.html
@@ -36,19 +36,19 @@
                                 </a>
                                 {% endif %}
                                 {% if not current_user.is_authenticated %}
-                                <a href="{{ url_for('main.login') }}" class="navbar-item">
+                                <a href="{{ url_for('auth.login') }}" class="navbar-item">
                                     Login
                                 </a>
-                                <a href="{{ url_for('main.signup') }}" class="navbar-item">
+                                <a href="{{ url_for('auth.signup') }}" class="navbar-item">
                                     Sign Up
                                 </a>
                                 {% endif %}
                                 {% if current_user.is_authenticated %}
-                                <a href="{{ url_for('main.logout') }}" class="navbar-item">
+                                <a href="{{ url_for('auth.logout') }}" class="navbar-item">
                                     Logout
                                 </a>
                                 {% endif %}
-                                <a href="{{ url_for('main.contact') }}" class="navbar-item">
+                                <a href="{{ url_for('auth.contact') }}" class="navbar-item">
                                     Contact
                                 </a>
                             </div>
