+++ project/templates/index.html
@@ -14,8 +14,8 @@
 </h3>
 <br>
 <h3>
-  Please <u><a href="{{ url_for('main.login') }}" >log in</a></u> or
-  <u><a href="{{ url_for('main.signup') }}" >sign up</a></u>.
+  Please <u><a href="{{ url_for('auth.login') }}" >log in</a></u> or
+  <u><a href="{{ url_for('auth.signup') }}" >sign up</a></u>.
 </h3>
 <br>
 <br>
